Add files via upload
This is the first version of a program that allows the user to enter in a string of numbers, specifically the words one and zero and it will convert said words into binary.

Any word that is not one or zero will be skipped over in the binary sequence. 

Planned fixes: 
Improve the program looping
limit the amount of words that can be entered. loop the program so that they can start over again if user exceeds the parameters